feat: Add Sass file for looping through header elements
This commit adds a new Sass file, 12-loop_header.scss, which contains a loop that generates CSS classes for each header element. Each generated class sets the font-size property to a different value based on the loop index. This will allow for dynamic and efficient styling of header elements throughout the codebase.