feat: Add Sass file for media query styling
This commit adds a new Sass file, 101-media_query.scss, which contains styling for a media query. The h1 element is styled with a font-size of 20px, and within a media query for screens with a maximum width of 320px, the font-size is set to 14px. This will allow for responsive font sizing on smaller screens.