```text refactor: Update 103-sort_strings.scss to use imported function for sorting
The code changes in this commit update the 103-sort_strings.scss file to use the imported function "sort" from the 103-sort_list.scss file for sorting the list of strings. This improves code organization and reusability by utilizing the existing function instead of defining a new one. The debug statement is also updated to use the sorted list variable.